+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [3.14.0] - 2022-02-09
+- **Added** app boolean option `setAllAttributes`
+- **Added** app callback option `onVdomUpdateElement`
+
 ## [3.13.10] - 2022-02-08
 - **Improvement** optimize esm;